SMA price oracle (#172)
* Add base SMA oracle implementation

This commit adds the base code for the simple moving average (SMA)
price oracle.

* Add stubs of overridden methods for `IOracleWRapper`

This commit adds stubs of two methods required by `IOracleWrapper`:

 - `fromWad`
 - `getPriceAndMetadata`

* Refactor contract properties

This commit clarifies some of the contract fields.

* Add setter for number of periods

* Add scaffold for SMA test suite

This commit sets up the test suite for the SMA price oracle.

* Mark `SMAOracle.setPeriods` as `public`

This commit marks the `setPeriods` method of `SMAOracle` as `public`.

* Make bounds check on periods inclusive

This commit corrects an errorneous bounds check on the number of
periods allowed in the `SMAOracle` contract.

* Add test cases for `SMAOracle.SMA`

This commit adds two test cases for the `SMAOracle.SMA` method. Both of
these test edge cases pertaining to the `k` parameter.

* Fix bounds check and loop bounds in SMA method

This commit fixes both the bounds check and the bounds on the `for`
loop in `SMAOracle.SMA`.

* Add test case for correct SMA calculation

This commit adds a test case for correct SMA calculation by
`SMAOracle.SMA`.

* Add test case for SMA periods matching dataset size

This commit adds a test case that covers the case of `k` being the same
size as the dataset itself.

* Add `updated` flag to `SMAOracle` contract

This commit adds a flag indicating whether the SMA oracle has been
initialised with historical price data yet.

* Add `onlyUpdated` modifier

This commit adds a modifier, `onlyUpdated`, in order to guard
operations that should only occur if the SMA oracle is in an
initialised state.

* Implement array left rotation

This commit adds a (pure) method, `SMAOracle.leftRotateWithPad`, which
rotates an array to the left by one element and sets the last element
of the array to a specified value.

* Add `IHistoricalOracleWrapper` interface

This commit adds a new interface in order to capture different oracle
functionality. This new interface allows querying past price data from
compliant oracles.

* Implement loading of historical prices

This commit implements the `initialise` function for `SMAOracle`. This
function loads historical price data from the underlying spot oracle in
order to initialise the observations array of the SMA oracle.

Note that this function name is not part of the common "Cloneable"
Solidity pattern.

* Implement SMA oracle update

This commit implements the `SMAOracle.update` method, which adds fresh
price data to the SMA oracle.

* Actually update SMA price on oracle update

This commit adds a recalculation of the reported SMA price whenever
`SMAOracle.update` is called.

* Implement `getPriceAndMetadata

This commit implements the `SMAOracle.getPriceAndMetadata` method.

* Remove configurable periods

This commit removes the ability to modify the number of SMA periods
post-deployment.

* Pick more realistic SMA periods

* Switch to observer model

* Add tests for price observer

* Add tests covering SMA oracle update

* Add test to check for array rotation on update

* Add array rotation to price observer

* Add WAD scaling

* Enable WAD support

* Add additional price observer tests

* Fix bugs and tests

* Fix bugs and tests

* Remove unnecessary commenting

* Update contracts/implementation/SMAOracle.sol

Co-authored-by: my_name_uhh <benpatch@hotmail.com>

* Add polling to IOracleWrapper

* Make SMAOracle compliant with new interface

* Add NatSpec comment to toWad

* Add NatSpec comment to fromWad

* Add missing deployments for dependent contracts

* Add writer to PriceObserver

* Avoid storing SMA price

* Add update interval checks and fix test suite deployments

* Add writer setter

* Permission PriceObserver.add

* Revert "Merge remote-tracking branch 'origin/pools-v2' into sma-oracle"

This reverts commit b57fd0c3976e8620d2c9f1fd3d2811c846a48351, reversing
changes made to 9501ebd48ab75849fbc85874e618c4ea4811e1c1.

* Update contracts/implementation/PriceObserver.sol

Co-authored-by: my_name_uhh <benpatch@hotmail.com>

* Update contracts/implementation/SMAOracle.sol

Co-authored-by: my_name_uhh <benpatch@hotmail.com>

* Update contracts/implementation/PriceObserver.sol

Co-authored-by: my_name_uhh <benpatch@hotmail.com>

* Update contracts/implementation/ChainlinkOracleWrapper.sol

Co-authored-by: my_name_uhh <benpatch@hotmail.com>

* Fix formatting

Co-authored-by: mynameuhh <my_name_uhh@protonmail.com>
Co-authored-by: my_name_uhh <benpatch@hotmail.com>