Utility scripts (#457)
* Add utility script for finding untested functions

This commit adds a utility script for finding (potentially) untested
functions in a given smart contract. It does this by diffing the list
of function names against the list of individual test files in the
relevant test subdirectory for the provided contract.

* Add utility script for extracting coverage reports

This commit adds a utility script for extracting human-readable
code coverage report tables from the existing code coverage framework.

* Add utility script for line counts

This commit adds a utility script for producing line count statistics.
This is useful generally but also specifically for audit engagements
also.